refactor(dat): extract Standard mode to DatStandardComposer, complete decomposition (#611 PR4)
Move ComposeStandard, StandardRowValues, StandardTextPath, ComposeProfile,
GenerateSimulatedHashes, and Standard header building from DatComposer to
DatStandardComposer. DatComposer is now an 80-line façade that delegates
to DatProductionComposer, DatLoadfileOnlyComposer, and DatStandardComposer.
DatComposer.cs 431 to 80 lines. DatStandardComposer.cs 389 lines. All DAT
files under 400 LOC. No behavior change.

Closes #611
